+++ src/main/kotlin/app/revanced/patches/youtube/layout/hide/mixplaylists/annotations/MixPlaylistsPatchCompatibility.kt
@@ -6,14 +6,15 @@ import app.revanced.patcher.annotation.Package
 @Compatibility(
     [Package(
         "com.google.android.youtube", arrayOf(
-            "17.49.37", 
+            "17.49.37",
             "18.03.36",
             "18.03.42",
             "18.04.35",
             "18.04.41",
             "18.05.32",
             "18.05.35",
-            "18.05.40"
+            "18.05.40",
+            "18.08.37"
         )
     )]
 )