+++ src/main/kotlin/app/revanced/patches/youtube/layout/spoofappversion/bytecode/patch/SpoofAppVersionPatch.kt
@@ -30,18 +30,17 @@ class SpoofAppVersionPatch : BytecodePatch(
         SpoofAppVersionFingerprint
     )
 ) {
-    companion object {
-        const val INTEGRATIONS_CLASS_DESCRIPTOR = "Lapp/revanced/integrations/patches/SpoofAppVersionPatch"
-    }
-
     override fun execute(context: BytecodeContext): PatchResult {
         SettingsPatch.PreferenceScreen.LAYOUT.addPreferences(
             SwitchPreference(
                 "revanced_spoof_app_version",
                 StringResource("revanced_spoof_app_version_title", "Spoof app version"),
                 false,
-                StringResource("revanced_spoof_app_version_summary_on", "Version spoofed to 17.30.34. If switched off, the old UI layout may remain until logging out or clearing app data"),
-                StringResource("revanced_spoof_app_version_summary_off", "Version not spoofed")
+                StringResource("revanced_spoof_app_version_summary_on", "Version spoofed to 17.30.34"),
+                StringResource("revanced_spoof_app_version_summary_off", "Version not spoofed"),
+                StringResource("revanced_spoof_app_version_user_dialog_message",
+                "App version will be spoofed to 17.30.34. This will give the old UI layout, but unknown side effects may occur."
+                        + " If later turned off, the old UI layout may remain until you log out or clear the app data.")
             )
         )
 
@@ -60,5 +59,9 @@ class SpoofAppVersionPatch : BytecodePatch(
         } ?: return SpoofAppVersionFingerprint.toErrorResult()
 
         return PatchResultSuccess()
+    }
+
+    private companion object {
+        const val INTEGRATIONS_CLASS_DESCRIPTOR = "Lapp/revanced/integrations/patches/SpoofAppVersionPatch"
     }
 }