+++ src/main/kotlin/app/revanced/patches/youtube/misc/debugging/patch/DebuggingPatch.kt
@@ -11,15 +11,16 @@ import app.revanced.patches.shared.settings.preference.impl.StringResource
 import app.revanced.patches.shared.settings.preference.impl.SwitchPreference
 import app.revanced.patches.youtube.misc.debugging.annotations.DebuggingCompatibility
 import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
+import app.revanced.patches.all.misc.debugging.patch.EnableAndroidDebuggingPatch
 import app.revanced.patches.youtube.misc.settings.bytecode.patch.SettingsPatch
 import org.w3c.dom.Element
 
 @Patch
-@Name("debugging")
-@DependsOn([IntegrationsPatch::class, SettingsPatch::class])
+@Name("enable-debugging")
+@DependsOn([IntegrationsPatch::class, SettingsPatch::class, EnableAndroidDebuggingPatch::class])
 @Description("Adds debugging options.")
 @DebuggingCompatibility
-@Version("0.0.1")
+@Version("0.0.2")
 class DebuggingPatch : ResourcePatch {
     override fun execute(context: ResourceContext): PatchResult {
         SettingsPatch.PreferenceScreen.MISC.addPreferences(
@@ -52,36 +53,18 @@ class DebuggingPatch : ResourcePatch {
                         ),
                         true,
                         StringResource("revanced_debug_toast_on_error_summary_on", "Toast shown if error occurs"),
-                        StringResource("revanced_debug_toast_on_error_summary_off", "Toast not shown if error occurs")
+                        StringResource("revanced_debug_toast_on_error_summary_off", "Toast not shown if error occurs"),
+                        StringResource("revanced_debug_toast_on_error_user_dialog_message",
+                            "Turning off error toasts hides all ReVanced error notifications." +
+                                    " This includes hiding normal network connection timeouts, " +
+                                    "but also hides notification of any unexpected and more serious errors."
+                        )
                     ),
                 ),
                 StringResource("revanced_debug_summary", "Enable or disable debugging options")
             )
         )
 
-        if (debuggable == true) {
-            context.xmlEditor["AndroidManifest.xml"].use { dom ->
-                val applicationNode = dom
-                    .file
-                    .getElementsByTagName("application")
-                    .item(0) as Element
-
-                // set application as debuggable
-                applicationNode.setAttribute("android:debuggable", "true")
-            }
-        }
-
         return PatchResultSuccess()
-    }
-
-    companion object : OptionsContainer() {
-        var debuggable: Boolean? by option(
-            PatchOption.BooleanOption(
-                key = "debuggable",
-                default = false,
-                title = "App debugging",
-                description = "Whether to make the app debuggable on Android.",
-            )
-        )
     }
 }