+++ src/main/kotlin/app/revanced/patches/youtube/misc/litho/filter/fingerprints/ComponentContextParserFingerprint.kt
@@ -5,13 +5,6 @@ import org.jf.dexlib2.Opcode
 
 object ComponentContextParserFingerprint : MethodFingerprint(
     opcodes = listOf(
-        Opcode.INVOKE_INTERFACE,
-        Opcode.MOVE_RESULT_OBJECT,
-        Opcode.CHECK_CAST,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.GOTO,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.CONST_16,
         Opcode.INVOKE_VIRTUAL,
         Opcode.IPUT_OBJECT,
         Opcode.NEW_INSTANCE